Fix autonumber on recent versions of Sphinx
Apparently, when the node created by the role function was inserted in
the tree, the `autonumber` class was replaced by a `Text` class, I wish
I knew why. But it seems this does not happen when using a different
base class for `autonumber`. Not sure if this is a real fix, but at
least it should be a workaround.

resolves #1